Test against current Java versions
Testing against more than one version of Java allows for ensuring
that there aren't any regressions in versions of Java that may be
used by end usres.

Java 8 is chosen because it is still fairly common and seems to be
what was used primarily in the past. Java 11 is the currently-supported
version.