Fix outdated tests and in-repo testing script
* Add main() wrappers for tests that were missing them
* Add support for standard input in tests (e.g., "sum.in" for "sum.lj")
* Skip running tests with no main()
* Ignore blank lines when checking output

Converting these to TravisCI tests is future work.

Note that there are still five tests that are known to be failing and are in the `failing` subfolder of the tests. For now it is ok for new contributions to be merged with these tests failing. Documenting and fixing these cases is future work.